[REFACTOR] 상세정보 페이지 관련 컴포넌트 폴더 분리
